app/requirements.txt
python-telegram-bot==13.2
requests
python-dotenv
bs4
pymongo
